Separate quota caching and request deduplication
