+++ src/main/java/module-info.java
@@ -2,6 +2,7 @@ module org.cryptomator.cloudaccess {
 	exports org.cryptomator.cloudaccess;
 	exports org.cryptomator.cloudaccess.api;
 	exports org.cryptomator.cloudaccess.api.exceptions;
+	exports org.cryptomator.cloudaccess.requestdecorator;
 
 	requires java.xml;
 	requires com.google.common;
